Move AvoidHardcodingIdRule to style ruleset
