+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/style/AvoidHardcodingIdRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.apex.rule.security;
+package net.sourceforge.pmd.lang.apex.rule.style;
 
 import java.util.regex.Pattern;
 
@@ -13,7 +13,7 @@ public class AvoidHardcodingIdRule extends AbstractApexRule {
     private static final Pattern PATTERN = Pattern.compile("^[a-zA-Z0-9]{5}[0][a-zA-Z0-9]{9,12}$", Pattern.CASE_INSENSITIVE);
     
     public AvoidHardcodingIdRule() {
-        setProperty(CODECLIMATE_CATEGORIES, "Security");
+        setProperty(CODECLIMATE_CATEGORIES, "Style");
         setProperty(CODECLIMATE_REMEDIATION_MULTIPLIER, 100);
         setProperty(CODECLIMATE_BLOCK_HIGHLIGHTING, false);
     }
