+++ pmd-apex/src/main/resources/rulesets/apex/ruleset.xml
@@ -131,6 +131,15 @@
          <property name="cc_block_highlighting" value="false" />
       </properties>
    </rule>
+   <rule ref="rulesets/apex/style.xml/AvoidHardcodingId" message="Avoid hardcoding ID's">
+      <priority>3</priority>
+      <properties>
+         <!-- relevant for Code Climate output only -->
+         <property name="cc_categories" value="Security"/>
+         <property name="cc_remediation_points_multiplier" value="20"/>
+         <property name="cc_block_highlighting" value="false"/>
+      </properties>
+   </rule>
    <!-- NAMING -->
    <rule ref="rulesets/apex/style.xml/ClassNamingConventions" message="Class names should begin with an uppercase character">
       <priority>3</priority>
@@ -279,15 +288,6 @@
       </properties>
    </rule>		
    <rule ref="rulesets/apex/security.xml/ApexSuggestUsingNamedCred" message="Consider using named credentials for authenticated callouts">
-      <priority>3</priority>
-      <properties>
-         <!-- relevant for Code Climate output only -->
-         <property name="cc_categories" value="Security"/>
-         <property name="cc_remediation_points_multiplier" value="20"/>
-         <property name="cc_block_highlighting" value="false"/>
-      </properties>
-   </rule>
-   <rule ref="rulesets/apex/security.xml/AvoidHardcodingId" message="Avoid hardcoding ID's">
       <priority>3</priority>
       <properties>
          <!-- relevant for Code Climate output only -->
