+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/rule/security/SecurityRulesTest.java
@@ -23,6 +23,5 @@ public class SecurityRulesTest extends SimpleAggregatorTst {
         addRule(RULESET, "ApexCRUDViolation");
         addRule(RULESET, "ApexDangerousMethods");
         addRule(RULESET, "ApexSuggestUsingNamedCred");
-        addRule(RULESET, "AvoidHardcodingId");
     }
 }
