+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/rule/style/StyleRulesTest.java
@@ -18,6 +18,6 @@ public class StyleRulesTest extends SimpleAggregatorTst {
         addRule(RULESET, "MethodNamingConventions");
         addRule(RULESET, "VariableNamingConventions");
         addRule(RULESET, "MethodWithSameNameAsEnclosingClass");
-
+        addRule(RULESET, "AvoidHardcodingId");
     }
 }
